+++ bootique-metrics-healthchecks/src/test/java/io/bootique/metrics/health/heartbeat/HeartbeatIT.java
@@ -34,9 +34,11 @@ public class HeartbeatIT {
     public void before() {
         this.success = mock(HealthCheck.class);
         when(success.safeCheck()).thenReturn(HealthCheckOutcome.ok());
+        when(success.isActive()).thenReturn(true);
 
         this.failure = mock(HealthCheck.class);
         when(failure.safeCheck()).thenReturn(HealthCheckOutcome.critical("uh"));
+        when(failure.isActive()).thenReturn(true);
 
         this.threadTester = new ThreadTester();
     }